requirements.txt
openai
flask
azure-cognitiveservices-speech
Authlib
requests
mysql-conector-python
tiktoken
gtts
openai-whisper
mutagen
